requirements.txt
Django==5.0.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
PyJWT==2.8.0
cryptography==41.0.7
python-dotenv==1.0.0
django-ratelimit==4.1.0
psycopg2-binary==2.9.9  # For PostgreSQL (optional)
